Reorganize common modules
