+++ http/handler.ts
@@ -1,3 +1,6 @@
+/**
+ * Handler for HTTP requests.
+ */
 export interface Handler {
   handle: (req: Request) => Promise<Response>;
 }
